Adds a MatchBodyModifier and some tests
When http vcr tries to match a request, it uses a `Matcher`. Our custom
Matcher for opensearch uses the request body to ensure that the bodies
of the requests are the same. However, when time, or a random value is
included in the body, the matcher will not work after the initial
recording since those time or random values will not be the same on
subsequent runs.

To solve this issue, the custom Matcher accepts a MatchBodyModifier
which allows for the request body to be modified prior to checking their
equality. This gives the opportunity for test creators to unmarshal the
body and remove time and random values before the equality check.

 - Adds an example MatchBodyModifier to the go-lanai test
 - Also adds comments to the matcher modifier and the test